pyproject.toml
[tool.poetry]
name = "evanduke"
version = "1.0.0-rc6"
description = ""
authors = ["Heather Duke <heather@heathernoelduke.com>"]
readme = "README.md"
include = [
    "enterprises/**/*",
    "evanduke/**/*",
    "manage.py",
    "pyproject.toml",
    "poetry.lock",
]

[tool.poetry.dependencies]
python = "^3.11"
gunicorn = "^20.1.0"
django = "^4.2"
psycopg2-binary = "^2.9.6"
django-environ = "^0.10.0"
google-cloud-recaptcha-enterprise = "^1.12.1"
sentry-sdk = "^1.30.0"

[tool.poetry.group.dev.dependencies]
isort = "^5"
flake8 = "^6"
black = "^23"
pytest-django = "^4.5"
pytest-mock = "^3.11"
pytest-cov = "^4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
extend-exclude = "migrations/*"
line-length = 120

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "evanduke.settings"
python_files = ["*/tests.py", "*/test_*.py"]

[tool.isort]
profile = "black"
line_length = 120
